refactor: fix findbugs.

+++ ratis-server/src/main/java/org/apache/ratis/statemachine/impl/SimpleStateMachineStorage.java
@@ -124,20 +124,20 @@ public class SimpleStateMachineStorage implements StateMachineStorage {
           .map(FileInfo::getPath)
           .forEach(snapshotPath -> {
             LOG.info("Deleting old snapshot at {}", snapshotPath.toAbsolutePath());
-            final boolean deleted = FileUtils.deletePathQuietly(snapshotPath);
-            if (deleted) {
-              final File md5file = MD5FileUtil.getDigestFileForFile(snapshotPath.toFile());
-              FileUtils.deleteFileQuietly(md5file);
-            }
+            FileUtils.deletePathQuietly(snapshotPath);
           });
       // clean up the md5 files if the corresponding snapshot file does not exist
       try (DirectoryStream<Path> stream = Files.newDirectoryStream(stateMachineDir.toPath(),
           SNAPSHOT_MD5_FILTER)) {
         for (Path md5path : stream) {
-          final String md5file = md5path.getFileName().toString();
-          final File snapshotFile = new File(md5path.getParent().toFile(),
-              md5file.substring(0, md5file.length() - MD5_SUFFIX.length()));
-          if (!snapshotFile.exists()){
+          Path md5FileNamePath = md5path.getFileName();
+          if (md5FileNamePath == null) {
+            continue;
+          }
+          final String md5FileName = md5FileNamePath.toString();
+          final File snapshotFile = new File(stateMachineDir,
+              md5FileName.substring(0, md5FileName.length() - MD5_SUFFIX.length()));
+          if (!snapshotFile.exists()) {
             FileUtils.deletePathQuietly(md5path);
           }
         }
